Relocate abi encoding for `EigenDACert` (#11)
* relocate encoding

* relocate `CheckDACertStatus`

* reduce alloy imports

* remove unwraps

* remove comments

* add partial eq to enum

* rename file

* Update crate

---------
