+++ Cargo.toml
@@ -58,5 +58,7 @@ dotenv = "0.15.0"
 serial_test = "3.1.1"
 bincode = "1.3.3"
 ethers = "2.0"
-eigensdk = { git = "https://github.com/lambdaclass/eigen-rs", branch = "chore/bump-alloy-1.0" }
+eigensdk = "2.0.0"
 alloy = "1.0.0"
+alloy-primitives = "1.0.0"
+alloy-sol-types = "1.0.0"
