+++ crates/rust-eigenda-v2-client/src/payload_disperser.rs
@@ -4,21 +4,54 @@ use alloy::primitives::{Address, FixedBytes};
 use ark_bn254::G1Affine;
 use ark_ff::{BigInteger, PrimeField};
 use eigensdk::client_avsregistry::reader::{AvsRegistryChainReader, AvsRegistryReader};
-use rust_eigenda_v2_common::{EigenDACert, NonSignerStakesAndSignature, Payload, PayloadForm};
+use rust_eigenda_v2_common::{
+    BatchHeaderV2, EigenDACert, NonSignerStakesAndSignature, Payload, PayloadForm, SignedBatch,
+};
 use tiny_keccak::{Hasher, Keccak};
 
 use crate::{
     cert_verifier::CertVerifier,
-    core::{
-        eigenda_cert::{build_cert_from_reply, SignedBatch},
-        BlobKey,
-    },
+    core::{eigenda_cert::build_cert_from_reply, BlobKey},
     disperser_client::{DisperserClient, DisperserClientConfig},
     errors::{ConversionError, EigenClientError, PayloadDisperserError},
     generated::disperser::v2::{BlobStatus, BlobStatusReply, SignedBatch as SignedBatchProto},
     rust_eigenda_signers::{signers::private_key::Signer as PrivateKeySigner, Sign},
     utils::SecretUrl,
 };
+
+impl TryFrom<SignedBatchProto> for SignedBatch {
+    type Error = ConversionError;
+
+    fn try_from(value: SignedBatchProto) -> Result<Self, Self::Error> {
+        let header = match value.header {
+            Some(header) => BatchHeaderV2 {
+                batch_root: header.batch_root.try_into().map_err(|_| {
+                    ConversionError::SignedBatch("Failed parsing batch root".to_string())
+                })?,
+                reference_block_number: header.reference_block_number.try_into().map_err(|_| {
+                    ConversionError::SignedBatch(
+                        "Failed parsing reference block number".to_string(),
+                    )
+                })?,
+            },
+            None => return Err(ConversionError::SignedBatch("Header is None".to_string())),
+        };
+
+        let attestation = match value.attestation {
+            Some(value) => value.try_into()?,
+            None => {
+                return Err(ConversionError::SignedBatch(
+                    "Attestation is None".to_string(),
+                ))
+            }
+        };
+
+        Ok(Self {
+            header,
+            attestation,
+        })
+    }
+}
 
 #[derive(Clone, Debug)]
 pub struct PayloadDisperserConfig {
