+++ crates/rust-eigenda-v2-common/src/core.rs
@@ -1,8 +1,10 @@
 mod blob;
+mod check_da_cert_status;
 mod encoded_payload;
 mod payload;
 
 pub use blob::Blob;
+pub use check_da_cert_status::CheckDACertStatus;
 pub use encoded_payload::EncodedPayload;
 pub use payload::Payload;
