+++ crates/rust-eigenda-v2-common/src/lib.rs
@@ -3,7 +3,7 @@ mod eigenda_cert;
 mod errors;
 mod utils;
 
-pub use core::{Blob, EncodedPayload, Payload, PayloadForm};
+pub use core::{Blob, CheckDACertStatus, EncodedPayload, Payload, PayloadForm};
 pub use eigenda_cert::*;
 pub use errors::*;
 pub use utils::*;
